+++ jest.config.js
@@ -5,9 +5,6 @@ module.exports = {
         "js",
     ],
     preset: 'ts-jest',
-    setupFilesAfterEnv: [
-        './jest.setup.js'
-    ],
     testEnvironment: 'node',
     testMatch: [
         "**/*.test.ts"