+++ src/inversion-of-control/container-manager.ts
@@ -14,5 +14,6 @@ polarisContainer.bind(POLARIS_TYPES.GraphQLServer).to(PolarisGraphQLServer);
 polarisContainer.bind(POLARIS_TYPES.GraphQLLogger).to(PolarisGraphQLLogger);
 polarisContainer.bind(POLARIS_TYPES.Middleware).to(PolarisMiddleware);
 polarisContainer.bind(POLARIS_TYPES.Middleware).to(IrrelevantEntitiesMiddleware);
-polarisContainer.bind(POLARIS_TYPES.RealitiesHolderValidator).to(RealitiesHolderValidator);
+polarisContainer.bind(POLARIS_TYPES.SoftDeleteConfiguration).toConstantValue({});
 polarisContainer.bind(POLARIS_TYPES.RealitiesHolder).toConstantValue({});
+polarisContainer.bind(POLARIS_TYPES.RealitiesHolderValidator).to(RealitiesHolderValidator);