+++ test/integration-tests/test-server/schema/types/output/author.ts
@@ -1,0 +1,9 @@
+export const Author = `type Author implements CommonEntity {
+                        id: ID!
+                        firstName: String,
+                        lastName: String,
+                        creationDate: Date,
+                        lastUpdateDate: Date,
+                        dataVersion: Int!,
+                        realityId: Int!
+                     }`;