+++ test/integration-tests/test-server/schema/types/output/book.ts
@@ -1,11 +1,10 @@
 export const Book = `type Book implements CommonEntity {
                         testId: String,
                         id: ID!,
+                        title: String,
+                        author: Author,
                         creationDate: Date,
                         lastUpdateDate: Date,
                         dataVersion: Int!,
-                        title: String,
-                        author: String,
-                        otherBook: ID,
                         realityId: Int!
                      }`;