+++ test/integration-tests/tests/mutation.test.ts
@@ -1,5 +1,9 @@
-import { graphqlRequest } from '../test-server/client';
-import { finish, init } from '../test-server/run-test';
+import { getModelCreator } from '@enigmatis/mongo-driver';
+import { graphQLRequest } from '../test-server/client';
+import { authorSchema } from '../test-server/dal/author-model';
+import { startTestServer, stopTestServer } from '../test-server/run-test';
+import { Author } from '../test-server/schema/definitions/author';
+import { TestServer } from '../test-server/server';
 
 const headers = { 'reality-id': 1 };
 const createBookMutation = `mutation createBook ($book:BookInput!) {createBook(book:$book){testId}}`;
@@ -9,49 +13,51 @@ const updateBookMutation = `mutation updateBook ($bookId:String!, $update:Update
 const findBookQuery = `query bookById ($bookId:String!) {bookById(bookId:$bookId){title}}`;
 const deleteBookMutation = `mutation deleteBook ($bookId:String!) {
          deleteBook(bookId: $bookId){ title }}`;
-const defaultBookVariables = (title: string, testId: string) => ({ author: 'chen', title, testId });
+const defaultBookVariables = (title: string, testId: string) => ({ title, testId });
 
+let testServer: TestServer;
 beforeEach(() => {
-    return init();
+    testServer = new TestServer();
+    return startTestServer(testServer.server);
 });
 
 afterEach(() => {
-    return finish();
+    return stopTestServer(testServer.server);
 });
 
 describe('mutation tests', () => {
     test('create book, book is created ', async () => {
         const id = '1';
         const title = 'book';
-        await graphqlRequest(createBookMutation, headers, {
+        await graphQLRequest(createBookMutation, headers, {
             book: defaultBookVariables(title, id),
         });
-        const response: any = await graphqlRequest(findBookQuery, headers, { bookId: id });
+        const response: any = await graphQLRequest(findBookQuery, headers, { bookId: id });
         expect(response.bookById.title).toEqual(title);
-        await graphqlRequest(deleteBookMutation, headers, { bookId: id });
+        await graphQLRequest(deleteBookMutation, headers, { bookId: id });
     });
     test('update book, book is updated ', async () => {
         const id = '1';
-        await graphqlRequest(createBookMutation, headers, {
+        await graphQLRequest(createBookMutation, headers, {
             book: defaultBookVariables('book', id),
         });
         const updatedTitle = 'book updated';
-        await graphqlRequest(updateBookMutation, headers, {
+        await graphQLRequest(updateBookMutation, headers, {
             bookId: id,
             update: { title: updatedTitle },
         });
-        const response: any = await graphqlRequest(findBookQuery, headers, { bookId: id });
+        const response: any = await graphQLRequest(findBookQuery, headers, { bookId: id });
         expect(response.bookById.title).toEqual(updatedTitle);
-        await graphqlRequest(deleteBookMutation, headers, { bookId: id });
+        await graphQLRequest(deleteBookMutation, headers, { bookId: id });
     });
 
     test('delete book, book is deleted ', async () => {
         const id = '1';
-        await graphqlRequest(createBookMutation, headers, {
+        await graphQLRequest(createBookMutation, headers, {
             book: defaultBookVariables('book', id),
         });
-        await graphqlRequest(deleteBookMutation, headers, { bookId: id });
-        const response: any = await graphqlRequest(findBookQuery, headers, { bookId: id });
+        await graphQLRequest(deleteBookMutation, headers, { bookId: id });
+        const response: any = await graphQLRequest(findBookQuery, headers, { bookId: id });
         expect(response.bookById).toBeNull();
     });
 });